Adicionado o mínimo de orientação Agora coisas relacionadas a gráfico ficam em um diretório só
